requirements.txt
Django==5.2.3
djangorestframework==3.16.0
asgiref==3.8.1
sqlparse==0.5.3
python-dotenv==1.1.0